+++ packages/g-lite/src/css/interfaces.ts
@@ -131,6 +131,10 @@ export interface PropertyParseOptions {
   forceUpdateGeometry: boolean;
   usedAttributes: string[];
   memoize: boolean;
+  /**
+   * @experimental
+   */
+  skipDispatchAttrModifiedEvent?: boolean;
 }
 
 export interface StyleValueRegistry {