+++ packages/g-svg/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/g-svg
 
+## 2.1.1
+
+### Patch Changes
+
+- Updated dependencies [00d08e9]
+- Updated dependencies [8df267d]
+    - @antv/g-lite@2.7.0
+
 ## 2.1.0
 
 ### Minor Changes